+++ src/main/java/com/swmStrong/demo/domain/usageLog/dto/SaveUsageLogDto.java
@@ -5,6 +5,6 @@ public record SaveUsageLogDto(
         String app,
         String url,
         double duration,
-        Integer timestamp
+        double timestamp
 ) {
 }
