+++ src/main/java/com/swmStrong/demo/domain/usageLog/entity/UsageLog.java
@@ -15,7 +15,7 @@ public class UsageLog {
     private ObjectId id;
 
     private String userId;
-    private Integer timestamp;
+    private double timestamp;
     private double duration;
     private String app;
     private String title;
@@ -23,7 +23,7 @@ public class UsageLog {
     private ObjectId categoryId;
 
     @Builder
-    public UsageLog(String userId, Integer timestamp, double duration, String app, String title, String url, ObjectId categoryId) {
+    public UsageLog(String userId, double timestamp, double duration, String app, String title, String url, ObjectId categoryId) {
         this.userId = userId;
         this.timestamp = timestamp;
         this.duration = duration;
